move getTeachersList function to utils
